requirements.txt
asgiref==3.8.1
attrs==23.2.0
Brotli==1.1.0
certifi==2024.2.2
cffi==1.16.0
chardet==5.2.0
colorama==0.4.6
cssselect2==0.7.0
dj-database-url==2.1.0
Django==5.0.6
django-allauth==0.63.2
django-storages==1.14.3
djangorestframework==3.15.1
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==5.3.1
fonttools==4.52.4
gunicorn==22.0.0
h11==0.14.0
html5lib==1.1
idna==3.7
iniconfig==2.0.0
Jinja2==3.1.4
MarkupSafe==2.1.5
outcome==1.3.0.post0
packaging==24.0
pillow==10.3.0
pluggy==1.5.0
psycopg2==2.9.9
psycopg2-binary==2.9.9
pycparser==2.22
pydyf==0.10.0
PyJWT==1.7.1
pyphen==0.15.0
pypng==0.20220715.0
PySocks==1.7.1
pytest==8.2.1
pytest-django==4.8.0
python-decouple==3.8
python-dotenv==1.0.1
qrcode==7.4.2
reportlab==4.2.0
selenium==4.21.0
six==1.16.0
sniffio==1.3.1
sortedcontainers==2.4.0
sqlparse==0.5.0
tinycss2==1.3.0
trio==0.25.1
trio-websocket==0.11.1
typing_extensions==4.11.0
tzdata==2024.1
urllib3==2.2.1
webencodings==0.5.1
whitenoise==6.6.0
wsproto==1.2.0
zopfli==0.2.3
